+++ pkg/repositories/types.go
@@ -3,6 +3,7 @@ package repositories
 type Repo struct {
 	Name          string           `yaml:"name"`
 	ComponentType string           `yaml:"componentType"`
+	System        string           `yaml:"system"`
 	Gen           RepoGen          `yaml:"gen"`
 	Lifecycle     RepoLifecycle    `yaml:"lifecycle"`
 	Replacements  RepoReplacements `yaml:"replace"`