+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Export catalog data from static YAML files.
+- Accept `system` property when reading repository data.
 
 ## [0.1.0] - 2023-08-16