+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Export catalog data from static YAML files.
+- Accept `system` property when reading repository data.
+
 ### Removed
 
 - Removed debug logging on dependencies found.