+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-08-17
+
 ### Added
 
 - For components of type "service", we add a Grafana dashboard link.
@@ -80,7 +82,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code
 
-[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.9...v0.1.0
 [0.0.9]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.7...v0.0.8